+++ apion/src/main/scala/io/github/edadma/apion/StaticMiddleware.scala
@@ -47,7 +47,7 @@ object StaticMiddleware:
       if options.redirect && !urlPath.endsWith("/") then
         // Redirect to add trailing slash
         Future.successful(
-          Complete(Response(301, ResponseHeaders(Map("Location" -> s"$urlPath/")), "")),
+          Complete(Response(301, ResponseHeaders(Seq("Location" -> s"$urlPath/")), "")),
         )
       else if options.index then
         // Try to serve index.html
@@ -94,7 +94,7 @@ object StaticMiddleware:
         // Read and send file
         fs.readFile(path).toFuture
           .map { content =>
-            val headers = ResponseHeaders(Map(
+            val headers = ResponseHeaders(Seq(
               "Content-Type"   -> mimeType,
               "Content-Length" -> stats.size.toString,
               "Cache-Control"  -> s"max-age=${options.maxAge}",